Fix: move distribution to utils directory
